Refactor AddBoxForm.
